requirements.txt
matplotlib
pillow
torch
tqdm
transformers
datasets